Add script to build common platforms.
The idea is to be able to allow newbies to easily build images
for common hardware.  These images should be user-friendly, including
luci and other tools that may aid debugging and use of the platform.

Includes example that builds x86-64 with additional drivers and luci
so that it works nicely on apu2 and other generic hardware.

Signed-off-by: Ben Greear <greearb@candelatech.com>